1D Autocorrelation Wavelets Transform

The 1D AC Wavelet transform will take a 1D signal such as a single time series and decompose it into wavelet coefficients.

Orthogonal Filters

To begin, one must first specify the high pass and low pass autocorrelation filters to use in the decomposition. The types of filters available for use are Haar, Coiflet, Daubechies, Symlet, Battle, Beylkin, Vaidyanathan, and CDF. These filters are implemented within the Wavelets.jl package.

Forward AC Wavelet Transform

To perform the transform on a signal, use the acwt function.

AutocorrelationShell.ACTransforms.acwt — Function

acwt(x, wt[, L=maxtransformlevels(x)])

Perform a forward ac wavelet transform of the array x. This method works for the 2D case as well. The wavelet type wt determines the transform type. Refer to Wavelet.jl for a list of available methods.

Examples

acwt(x, wavelet(WT.db4))

See also: iacwt

Inverse AC Wavelet Transform

To reconstruct the signal from a array of AC Wavelet coefficients, use the iacwt function.

AutocorrelationShell.ACTransforms.iacwt — Function

iacwt

The inverse of acwt

See also: acwt