Auto-correlation Wavelets Utilities

AutocorrelationShell.jl contains functions that can be used for adding gaussian noise, computing signal to noise ratio, computing structural similarity index(SSIM), performing thresholding, visualizing matrix heat map, visualizing the decomposition of the 1D transform, or calculate the entropy of a vector.

Missing docstring.

Missing docstring for make_noisy(x::AbstractArray{<:Number}, rng::MersenneTwister, a::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for snr(f::AbstractArray{T}, g::AbstractArray{T}) where T<:Number. Check Documenter's build log for details.

Missing docstring.

Missing docstring for acthreshold(x::AbstractArray{<:Number}, type::AbstractString, t::Float64). Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_snr(y::AbstractArray{T, Integer}, x::AbstractArray{T, Integer}; type::AbstractString="hard", step::Float64=0.5) where T<:Number. Check Documenter's build log for details.

Missing docstring.

Missing docstring for get_ssim(y::AbstractArray{T}, x::AbstractArray{T}; type::AbstractString="hard", step::Float64=0.5) where T<:Number. Check Documenter's build log for details.

Missing docstring.

Missing docstring for acwt_heatmap(x::AbstractArray{<:Number}). Check Documenter's build log for details.

Missing docstring.

Missing docstring for wentropy(x::AbstractArray{T}, et::Wavelets.Entropy, nrm::T=norm(x)) where T<:Number. Check Documenter's build log for details.

Missing docstring.

Missing docstring for wiggle(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, EdgeColor=:black, FaceColor=:black, Overlap=true, Orient=:across, ZDir=:normal). Check Documenter's build log for details.

AutocorrelationShell.ACPlots.wiggle! — Method

wiggle!(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, ZDir=:normal)

Plot a set of shaded wiggles on the current displayed graphics

Arguments

wav::Array: matrix of waveform columns.

taxis::Array=1:size(wav,1): time axis vector

zaxis::Array=1:size(wav,2): space axis vector

sc::Float=1: scale factor/magnification.

EdgeColor::Symbol=:black: Sets edge of wiggles color.

FaceColor::Symbol=:black: Sets shading color of wiggles.

Overlap::bool=true: How signals are scaled. true - Signals overlap (default); false - Signals are scaled so they do not overlap.

Orient::Symbol=:across: Controls orientation of wiggles. :across - from left to right :down - from top to down

ZDir::Symbol=:normal: Direction of space axis. :normal - First signal at bottom (default) :reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC) Revised by Naoki Saito, Feb. 05, 2018