AutocorrelationShell.acwpt — Method

acwpt(x, P, Q)

Compute the autocorrelation wavelet packet transform for a given signal. Returns a binary tree object.

Arguments

x::Vector{<:Real}: 1 dimensional signal.

P::Vector{<:Real}: Low autocorrelation shell filter.

Q::Vector{<:Real}: High autocorrelation shell filter.

AutocorrelationShell.aciwpt — Method

aciwpt(tree)

Reconstructs the signal using the autocorrelation wavelet packet trasform bases.

Arguments

tree::BinaryNode: The root node of the wavelet packet tree (binary tree)

AutocorrelationShell.acwptBestBasisTree — Method

acwptBestBasisTree(node; direction, et)

Finds the best set of basis using a given entropy criterion.

Arguments

node::BinaryNode: Root node of the wavelet packet tree (binary)

direction::AbstractString: Direction(left or right) of the node relative to its parent. default: "right"

et::Wavelets.Entropy: Entropy criterion used to determine the best tree.