AutocorrelationShell.make_noisy — Method

make_noisy(x::AbstractArray{<:Number}, rng::MersenneTwister, a::Float64)

Adds gaussian white noise to an image.

Arguments

x::AbstractArray{<:Number}: image rng::MersenneTwister: random number generator ex) rng = MersenneTwister(123) a::Float64: A scaling parameter to adjust level of noise.

Example

using AutocorrelationShell, Images, FileIO, Random

# Load test image
img = load("./test/pictures/boat.jpg")
img = Float64.(Gray.(img))

rng = MersenneTwister(123)
noisy_image = make_noisy(img, rng, 0.7)

AutocorrelationShell.snr — Method

snr(f::AbstractArray{T}, g::AbstractArray{T}) where T<:Number

Computes the signal to noise ratio between two signals. Returns signal to noise ratio(dB).

Arguments

f::AbstractArray{<:Number}: Original signal g::AbstractArray{<:Number}: Noisy signal

AutocorrelationShell.acthreshold — Method

acthreshold(x::AbstractArray{<:Number}, type::AbstractString, t::Float64)

Thresholds a given array of coeffecients using either hard thresholding or soft thresholding.

Arguments

x::AbstractArray{<:Number}: Autocorrelation wavelet coefficient array

type::AbstractString: Threshold type. "hard" or "soft"

t::Float64: threshold value

AutocorrelationShell.get_snr — Method

get_snr(y::AbstractArray{T}, x::AbstractArray{T}; type::AbstractString="hard", step::Float64=0.5) where T<:Number

Thresholds the wavelet coefficients matrix of a noisy signal, reconstructs it, then computes the signal to noise ratio between the reconstructed signal and the original non-noisy signal. Returns a list of the ratio of non-zero coefficient at each threshold and a list containing the signal to noise ratio.

Arguments

y: original 2D signal

x: wavelet coefficient matrix of a noisy 2D signal

type: type of thresholding (soft or hard)

step: step at which to increase the threshold

Example

img = load("./test/pictures/lenna.jpg")
img = Float64.(Gray.(img))

# Add noise to image
rng = MersenneTwister(123)
noisy = make_noisy(img, rng, 0.7)

# Apply Wavelet transform
H = wavelet(WT.db2)
L = 2
Q = qfilter(H)
P = pfilter(H)

ac_noisy = ac2d(noisy, L, P, Q)

coef_ratio, snr_list = get_snr(img, ac_noisy, type="soft", step=0.5)

AutocorrelationShell.get_ssim — Method

get_ssim(y::AbstractArray{T}, x::AbstractArray{T}; type::AbstractString="hard", step::Float64=0.5) where T<:Number

Thresholds the wavelet coefficients matrix of a noisy signal, reconstructs it, then computes the structural similarity index between the reconstructed signal and the original non-noisy signal. Returns a list of the ratio of non-zero coefficient at each threshold and a list containing the SSIM values.

Arguments

y: original 2D signal

x: wavelet coefficient matrix of a noisy 2D signal

type: type of thresholding (soft or hard)

step: step at which to increase the threshold

Example

using AutocorrelationShell, Wavelets, FileIO, Images, ImageQualityIndexes, Random

img = load("./test/pictures/lenna.jpg")
img = Float64.(Gray.(img))

# Add noise to image
rng = MersenneTwister(123)
noisy = make_noisy(img, rng, 0.7)

# Apply Wavelet transform
H = wavelet(WT.db2)
L = 2
Q = qfilter(H)
P = pfilter(H)

ac_noisy = acwt2d(noisy; L=2, P=P, Q=Q)

coef_ratios, ssim_values = get_ssim(img, ac_noisy; type="hard", step=0.5)

AutocorrelationShell.acwt_heatmap — Method

acwt_heatmap(x::AbstractArray{<:Number})

Takes the natural log of the absolute value of each cell, and plots a heatmap. Can be used for visualizing an input image or the coefficient matrix of a 2D ACW decomposition.

Arguments

x::AbstractArray{<:Number}: A image or a 2D matrix

AutocorrelationShell.wiggle — Method

wiggle(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, ZDir=:normal)

Plot a set of shaded wiggles

Arguments

wav::Array: matrix of waveform columns.

taxis::Array=1:size(wav,1): time axis vector

zaxis::Array=1:size(wav,2): space axis vector

sc::Float=1: scale factor/magnification.

EdgeColor::Symbol=:black: Sets edge of wiggles color.

FaceColor::Symbol=:black: Sets shading color of wiggles.

Overlap::bool=true: How signals are scaled. true - Signals overlap (default); false - Signals are scaled so they do not overlap.

Orient::Symbol=:across: Controls orientation of wiggles. :across - from left to right :down - from top to down

ZDir::Symbol=:normal: Direction of space axis. :normal - First signal at bottom (default) :reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC) Revised by Naoki Saito, Feb. 05, 2018

AutocorrelationShell.wiggle! — Method

wiggle!(wav; taxis=1:size(wav,1), zaxis=1:size(wav,2), sc=1, EdgeColor=:black, FaceColor=:black, Orient=:across, Overlap=true, ZDir=:normal)

Plot a set of shaded wiggles on the current displayed graphics

Arguments

wav::Array: matrix of waveform columns.

taxis::Array=1:size(wav,1): time axis vector

zaxis::Array=1:size(wav,2): space axis vector

sc::Float=1: scale factor/magnification.

EdgeColor::Symbol=:black: Sets edge of wiggles color.

FaceColor::Symbol=:black: Sets shading color of wiggles.

Overlap::bool=true: How signals are scaled. true - Signals overlap (default); false - Signals are scaled so they do not overlap.

Orient::Symbol=:across: Controls orientation of wiggles. :across - from left to right :down - from top to down

ZDir::Symbol=:normal: Direction of space axis. :normal - First signal at bottom (default) :reverse - First signal at top.

Translated by Nicholas Hausch – MATLAB file provided by Naoki Saito The previous MATLAB version contributors are: Anthony K. Booer (SLB) and Bradley Marchand (NSWC-PC) Revised by Naoki Saito, Feb. 05, 2018